Add support for generic ModelTypes
Previously, when using a custom data fetcher for Glide, the first param of the RequestListener must match the data fetcher model type.

This change simply allows a generic type to be passed to the RequestListener, rather than forcing the String type, which means GlidePalette can work with custom data fetchers.